testing started on basic structure

+++ src/main/java/behaviours/ISell.java
@@ -2,5 +2,5 @@ package behaviours;
 
 public interface ISell {
 
-    public int calculateMarkUp();
+    public double calculateMarkUp();
 }
